cmplxs/cscalar: new package containing non-vector functions from cmplxs
