+++ cmplxs/cscalar/doc.go
@@ -1,0 +1,6 @@
+// Copyright ©2020 The Gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// Package cscalar provides a set of helper routines for dealing with complex128 values.
+package cscalar // import "gonum.org/v1/gonum/cmplxs/cscalar"
